move duplicate code to utils
